project_sourse_code/requirements.txt
Flask>=2.0
mysql-connector-python>=8.0
